# Conflicts:
#	core/pkg/opencost/opencost_codecs.go
#	core/pkg/util/buffer.go
#	core/pkg/util/bufferhelper.go
#	core/pkg/util/bufferpool.go
#	core/pkg/util/monitor/memory/helpers.go
#	core/pkg/util/monitor/memory/helpers_test.go
#	core/pkg/util/monitor/memory/memorylimitstats.go
#	core/pkg/util/monitor/memory/memorylimitstats_test.go
#	core/pkg/util/stringutil/lrubank.go
#	core/pkg/util/stringutil/stringutil_test.go

+++ core/pkg/util/bufferhelper.go
@@ -76,7 +76,7 @@ func readInt(r *bytes.Buffer, data *int) error {
 		return err
 	}
 
-	*data = int(order.Uint32(bs))
+	*data = int(int32(order.Uint32(bs)))
 	return nil
 }
 
@@ -246,7 +246,7 @@ func readBuffInt(r *bufio.Reader, data *int) error {
 		return err
 	}
 
-	*data = int(order.Uint32(bs))
+	*data = int(int32(order.Uint32(bs)))
 	return nil
 }
 
@@ -442,7 +442,7 @@ func writeInt(w *bytes.Buffer, data int) error {
 	var b [4]byte
 	bs := b[:]
 
-	binary.LittleEndian.PutUint32(bs, uint32(data))
+	binary.LittleEndian.PutUint32(bs, uint32(int32(data)))
 	_, err := w.Write(bs)
 	return err
 }
